refactor: 결과 판단 로직 이동 (Player -> Participants)
