+++ README.md
@@ -54,9 +54,7 @@
 - [x] Participants | 카드 합 구하는 로직 변경
 - [x] YesOrNo | Enum으로 합체
 - [x] Dealer | 더 뽑을 수 있는지 스스로 판단 (참여자 객체에서 판단하게 함)
-- [ ] Player | 승패 상태를 저장 위치 고민
 - [x] Deck | 카드를 명시적으로 넣어 주기 (이러면 Deck이 셔플 방법을 몰라도 됨)
-- [ ] ParticipantStatus | 승부를 벌이는 일은 누구의 역할인지 고민 (Status or Participant)
 - [x] Blackjack | shouldStopDrawing 메소드가 한 가지 일만 하도록 변경
 - [x] Participants | 가시성 변경자 수정
 - [x] 8명 인원 제한
@@ -78,7 +76,7 @@
 
 - [x] 패키지 구조 변경
 - [x] ActionType | Boolean -> String 타입 변경
-- [ ] Player | 결과 판단 로직 위치 고민 (Player vs Participant)
+- [x] Player | 결과 판단 로직 위치 고민 (Player vs Participant)
 - [x] Player & Dealer | getInitCard 메소드 로직 변경
 - [x] test | 메소드명 직관적으로 변경
 - [ ] Introduce | 이름 필수 요소 고민
