+++ src/main/kotlin/blackjack/domain/model/participant/Participants.kt
@@ -1,5 +1,6 @@
 package blackjack.domain.model.participant
 
+import blackjack.domain.model.GameResult
 import blackjack.domain.model.card.Card
 import blackjack.domain.model.card.Hand
 
@@ -13,6 +14,25 @@ abstract class Participants() {
     fun receiveCard(card: Card) {
         hand.append(card)
     }
+
+    fun compareScores(participant: Participants): GameResult =
+        when {
+            hand.isBust() -> GameResult.Lose
+            participant.hand.isBust() -> GameResult.Win
+            participant.hand.isBlackjack() && !hand.isBlackjack() -> GameResult.Lose
+            !participant.hand.isBlackjack() && hand.isBlackjack() -> GameResult.Win
+            else -> compareNumber(sumCardNumber, participant.sumCardNumber)
+        }
+
+    private fun compareNumber(
+        target: Int,
+        other: Int,
+    ): GameResult =
+        when {
+            target < other -> GameResult.Lose
+            target > other -> GameResult.Win
+            else -> GameResult.Draw
+        }
 
     abstract fun canHit(): Boolean
