+++ src/main/kotlin/blackjack/domain/model/participant/Player.kt
@@ -1,6 +1,5 @@
 package blackjack.domain.model.participant
 
-import blackjack.domain.model.GameResult
 import blackjack.domain.model.card.Card
 import blackjack.domain.model.card.Hand
 
@@ -14,23 +13,4 @@ data class Player(override val name: String) : Participants() {
     override fun getInitCard(): List<Card> {
         return hand.cards.take(2)
     }
-
-    fun compareScores(dealer: Dealer): GameResult =
-        when {
-            hand.isBust() -> GameResult.Lose
-            dealer.hand.isBust() -> GameResult.Win
-            dealer.hand.isBlackjack() && !hand.isBlackjack() -> GameResult.Lose
-            !dealer.hand.isBlackjack() && hand.isBlackjack() -> GameResult.Win
-            else -> compareNumber(sumCardNumber, dealer.sumCardNumber)
-        }
-
-    private fun compareNumber(
-        target: Int,
-        other: Int,
-    ): GameResult =
-        when {
-            target < other -> GameResult.Lose
-            target > other -> GameResult.Win
-            else -> GameResult.Draw
-        }
 }
